+++ crates/compiler/src/bytecode_assembler.rs
@@ -102,6 +102,14 @@ where
     parse_svnt(arg)
 }
 
+fn one_svnt_argument_allow_zero<T>(args: &[&str]) -> Result<T, ChannelError>
+where
+    T: SignedValueNewType,
+{
+    let arg = one_argument(args)?;
+    parse_svnt_allow_zero(arg)
+}
+
 fn one_uvnt_argument<T>(args: &[&str]) -> Result<T, ChannelError>
 where
     T: UnsignedValueNewType,
@@ -326,6 +334,31 @@ fn parse_i32(src: &str, missing_sign_err: &ValueError) -> Result<i32, ValueError
     }
 }
 
+fn parse_i32_allow_zero(src: &str, missing_sign_err: &ValueError) -> Result<i32, ValueError> {
+    if let Some(s) = src.strip_prefix("+$") {
+        match i32::from_str_radix(s, 16) {
+            Ok(i) => Ok(i),
+            Err(_) => Err(ValueError::CannotParseHex(src.to_owned())),
+        }
+    } else if let Some(s) = src.strip_prefix("-$") {
+        match i32::from_str_radix(s, 16) {
+            Ok(i) => Ok(-i),
+            Err(_) => Err(ValueError::CannotParseHex(src.to_owned())),
+        }
+    } else {
+        match src.bytes().next() {
+            Some(b'-') | Some(b'+') => match src.parse() {
+                Ok(i) => Ok(i),
+                Err(_) => Err(ValueError::CannotParseSigned(src.to_owned())),
+            },
+            _ => match src.parse::<i32>() {
+                Ok(0) => Ok(0),
+                _ => Err(missing_sign_err.clone()),
+            },
+        }
+    }
+}
+
 /// Parse UnsignedValueNewType
 fn parse_uvnt<T>(s: &str) -> Result<T, ChannelError>
 where
@@ -340,6 +373,14 @@ where
     T: SignedValueNewType,
 {
     Ok(parse_i32(s, &T::MISSING_SIGN_ERROR)?.try_into()?)
+}
+
+/// Parse SignedValueNewType
+fn parse_svnt_allow_zero<T>(s: &str) -> Result<T, ChannelError>
+where
+    T: SignedValueNewType,
+{
+    Ok(parse_i32_allow_zero(s, &T::MISSING_SIGN_ERROR)?.try_into()?)
 }
 
 pub fn parse_asm_line(bc: &mut Bytecode, line: &str) -> Result<(), ChannelError> {
@@ -441,6 +482,9 @@ pub fn parse_asm_line(bc: &mut Bytecode, line: &str) -> Result<(), ChannelError>
 
        disable_early_release 0 no_arguments,
        set_early_release 3 early_release_arguments,
+
+       set_detune 1 one_svnt_argument_allow_zero,
+       disable_detune 0 no_arguments,
 
        adjust_volume 1 one_svnt_argument,
        set_volume 1 one_uvnt_argument,