+++ crates/compiler/src/value_newtypes.rs
@@ -257,6 +257,50 @@ macro_rules! i8_value_newtype {
     };
 }
 
+macro_rules! i16_value_newtype {
+    ($name:ident, $range_error:ident, $missing_error:ident, $missing_sign_error:ident, $min:expr, $max:expr) => {
+        #[derive(Debug, Copy, Clone, PartialEq)]
+        pub struct $name(i16);
+
+        #[allow(dead_code)]
+        impl $name {
+            pub const MIN: Self = Self($min);
+            pub const MAX: Self = Self($max);
+
+            pub fn as_i16(&self) -> i16 {
+                self.0
+            }
+
+            pub fn is_negative(&self) -> bool {
+                self.0 < 0
+            }
+        }
+
+        impl SignedValueNewType for $name {
+            type ValueType = i16;
+
+            const MISSING_ERROR: ValueError = ValueError::$missing_error;
+            const MISSING_SIGN_ERROR: ValueError = ValueError::$missing_sign_error;
+
+            fn value(&self) -> Self::ValueType {
+                self.0
+            }
+        }
+
+        impl TryFrom<i32> for $name {
+            type Error = ValueError;
+
+            fn try_from(value: i32) -> Result<Self, Self::Error> {
+                if value >= Self::MIN.0.into() && value <= Self::MAX.0.into() {
+                    Ok($name(value.try_into().unwrap()))
+                } else {
+                    Err(ValueError::$range_error(value))
+                }
+            }
+        }
+    };
+}
+
 macro_rules! i16_non_zero_value_newtype {
     ($name:ident, $range_error:ident, $missing_error:ident, $missing_sign_error:ident, $zero_error:ident, $min:expr, $max:expr) => {
         #[derive(Debug, Copy, Clone, PartialEq)]
@@ -304,6 +348,6 @@ macro_rules! i16_non_zero_value_newtype {
 }
 
 pub(crate) use {
-    i16_non_zero_value_newtype, i8_value_newtype, u16_value_newtype, u8_0_is_256_value_newtype,
-    u8_value_newtype,
+    i16_non_zero_value_newtype, i16_value_newtype, i8_value_newtype, u16_value_newtype,
+    u8_0_is_256_value_newtype, u8_value_newtype,
 };