+++ crates/compiler/tests/mml.rs
@@ -810,6 +810,191 @@ fn test_early_release_comma_f0_gain_is_err() {
 }
 
 #[test]
+fn test_detune() {
+    assert_line_matches_bytecode("D+1 D0", &["set_detune +1", "disable_detune"]);
+    // Testing `set_detune 0` is allowed in bytecode assembly
+    assert_line_matches_bytecode("D+1 D0", &["set_detune +1", "set_detune 0"]);
+
+    assert_line_matches_bytecode("D+10", &["set_detune +10"]);
+    assert_line_matches_bytecode("D-10", &["set_detune -10"]);
+    assert_line_matches_bytecode("D+300", &["set_detune +300"]);
+    assert_line_matches_bytecode("D-300", &["set_detune -300"]);
+
+    assert_line_matches_bytecode_bytes(
+        "D+1 D0",
+        &[opcodes::SET_DETUNE_P8, 1, opcodes::SET_DETUNE_P8, 0],
+    );
+
+    assert_line_matches_bytecode_bytes("D+10", &[opcodes::SET_DETUNE_P8, 10]);
+    assert_line_matches_bytecode_bytes("D-10", &[opcodes::SET_DETUNE_N8, 0xf6]);
+    assert_line_matches_bytecode_bytes("D+300", &[opcodes::SET_DETUNE_I16, 0x2c, 0x01]);
+    assert_line_matches_bytecode_bytes("D-300", &[opcodes::SET_DETUNE_I16, 0xd4, 0xfe]);
+
+    // Test limits
+    assert_line_matches_bytecode("D+16383", &["set_detune +$3fff"]);
+    assert_line_matches_bytecode("D-16383", &["set_detune -$3fff"]);
+    assert_error_in_mml_line(
+        "D+16384",
+        1,
+        ValueError::DetuneValueOutOfRange(16384).into(),
+    );
+    assert_error_in_mml_line(
+        "D-16384",
+        1,
+        ValueError::DetuneValueOutOfRange(-16384).into(),
+    );
+
+    assert_error_in_mml_line("D", 1, ValueError::NoDetuneValue.into());
+    assert_error_in_mml_line("D8", 1, ValueError::NoDetuneValueSign.into());
+}
+
+#[test]
+fn test_detune_deduplication() {
+    // Channel starts with detune set to 0
+    assert_line_matches_line("D0 D0 D0", "");
+
+    assert_line_matches_line(
+        "D0 D+20 D+20 D+21 D-15 D-15 D+5 D+20 D0",
+        "D0 D+20      D+21 D-15      D+5 D+20 D0",
+    );
+
+    assert_line_matches_bytecode(
+        "D+25 [a]5 D-30",
+        &[
+            "set_detune +25",
+            "start_loop",
+            "play_note a4 24",
+            "end_loop 5",
+            "set_detune -30",
+        ],
+    );
+
+    assert_line_matches_bytecode(
+        "D+25 [a]5 D+25",
+        &[
+            "set_detune +25",
+            "start_loop",
+            "play_note a4 24",
+            "end_loop 5",
+            // D+25 deduplicated
+        ],
+    );
+
+    assert_line_matches_bytecode(
+        "D+10 a [D+25 b]5 D+25",
+        &[
+            "set_detune +10",
+            "play_note a4 24",
+            "start_loop",
+            "set_detune +25",
+            "play_note b4 24",
+            "end_loop 5",
+            // D+25 deduplicated
+        ],
+    );
+
+    assert_line_matches_bytecode(
+        "D+300 [D+300 a : D-40 b]5 D-40",
+        &[
+            "set_detune +300",
+            "start_loop",
+            "set_detune +300",
+            "play_note a4 24",
+            "skip_last_loop",
+            "set_detune -40",
+            "play_note b4 24",
+            "end_loop 5",
+            "set_detune -40",
+        ],
+    );
+
+    assert_line_matches_bytecode(
+        "D+300 [D+300 a : D-40 b]5 D+300",
+        &[
+            "set_detune +300",
+            "start_loop",
+            "set_detune +300",
+            "play_note a4 24",
+            "skip_last_loop",
+            "set_detune -40",
+            "play_note b4 24",
+            "end_loop 5",
+            // D+300 deduplicated
+        ],
+    );
+
+    assert_mml_channel_a_matches_bytecode(
+        r##"
+@1 dummy_instrument
+
+!s D+10 a
+
+A @1 D+20 !s D+20 a
+"##,
+        &[
+            "set_instrument dummy_instrument",
+            "set_detune +20",
+            "call_subroutine s",
+            "set_detune +20",
+            "play_note a4 24",
+        ],
+    );
+
+    assert_mml_channel_a_matches_bytecode(
+        r##"
+@1 dummy_instrument
+
+!s D+10 a
+
+A @1 D+20 !s D+10 a
+"##,
+        &[
+            "set_instrument dummy_instrument",
+            "set_detune +20",
+            "call_subroutine s",
+            // D+10 deduplicated
+            "play_note a4 24",
+        ],
+    );
+
+    // Also tests D0 in a subroutine is not deduplicated
+    assert_mml_channel_a_matches_bytecode(
+        r##"
+@1 dummy_instrument
+
+!s D0 a
+
+A @1 D+10 !s D0 a
+"##,
+        &[
+            "set_instrument dummy_instrument",
+            "set_detune +10",
+            "call_subroutine s",
+            // D0 deduplicated
+            "play_note a4 24",
+        ],
+    );
+
+    // Test call subroutine does not change detune
+    assert_mml_channel_a_matches_bytecode(
+        r##"
+@1 dummy_instrument
+
+!s a
+
+A @1 D+30 !s D+30 a
+"##,
+        &[
+            "set_instrument dummy_instrument",
+            "set_detune +30",
+            "call_subroutine s",
+            // D+30 deduplicated
+            "play_note a4 24",
+        ],
+    );
+}
+
+#[test]
 fn play_long_note() {
     // `wait` can rest for 1 to 256 ticks.
     // `rest` can rest for 2 to 257 tick.